requirements.txt
streamlit
python-dotenv
groq
sentence-transformers
scikit-learn
pypdf
nbformat
numpy==1.26.4
